Extract referencesData collection function
